refactor: centralize remaining TransactionContext impls from `mls` module [WPB-15535]
